Split Commands.py so that each command has a file.
